Refactor metrics definitions to be easily understandable
As part of #9209 in preparation for adding / modifying telemetry to
comply with the convention set architecture, move all the definitions
into one place so we can see exactly what metrics are are exposing and
what their structure is.

Not quite sure that the initialization of the peformance based metrics
actually needs to be done separately from the rest, but keeping it that
way for now to be consistent.